// src/storefront/generated/schema.ts
export type Scalars = {
    Boolean: boolean,
    Color: any,
    DateTime: any,
    Decimal: any,
    Float: number,
    HTML: any,
    ID: string,
    Int: number,
    JSON: any,
    String: string,
    URL: any,
    UnsignedInt64: any,
}
// ...
export interface Attribute {
    /** Key or name of the attribute. */
    key: Scalars['String']
    /** Value of the attribute. */
    value: (Scalars['String'] | null)
    __typename: 'Attribute'
}
// ...
export type BaseCartLine = (CartLine) & { __isUnion?: true }


/**
 * An auto-generated type for paginating through multiple BaseCartLines.
 * 
 */
export interface BaseCartLineConnection {
    /** A list of edges. */
    edges: BaseCartLineEdge[]
    /** A list of the nodes contained in BaseCartLineEdge. */
    nodes: BaseCartLine[]
    /** Information to aid in pagination. */
    pageInfo: PageInfo
    __typename: 'BaseCartLineConnection'
}
// ...
export interface Cart {
    /** An attribute associated with the cart. */
    attribute: (Attribute | null)
    /** The attributes associated with the cart. Attributes are represented as key-value pairs. */
    attributes: Attribute[]
    /** Information about the buyer that is interacting with the cart. */
    buyerIdentity: CartBuyerIdentity
    /** The URL of the checkout for the cart. */
    checkoutUrl: Scalars['URL']
    /** The estimated costs that the buyer will pay at checkout. The costs are subject to change and changes will be reflected at checkout. The `cost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing). */
    cost: CartCost
    /** The date and time when the cart was created. */
    createdAt: Scalars['DateTime']
    /**
     * The delivery groups available for the cart, based on the buyer identity default
     * delivery address preference or the default address of the logged-in customer.
     * 
     */
    deliveryGroups: CartDeliveryGroupConnection
    /** The discounts that have been applied to the entire cart. */
    discountAllocations: CartDiscountAllocation[]
    /**
     * The case-insensitive discount codes that the customer added at checkout.
     * 
     */
    discountCodes: CartDiscountCode[]
    /**
     * @deprecated Use `cost` instead.
     * The estimated costs that the buyer will pay at checkout.
     * The estimated costs are subject to change and changes will be reflected at checkout.
     * The `estimatedCost` field uses the `buyerIdentity` field to determine
     * [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing).
     * 
     */
    estimatedCost: CartEstimatedCost
    /** A globally-unique ID. */
    id: Scalars['ID']
    /** A list of lines containing information about the items the customer intends to purchase. */
    lines: BaseCartLineConnection
    /** Returns a metafield found by namespace and key. */
    metafield: (Metafield | null)
    /**
     * The metafields associated with the resource matching the supplied list of namespaces and keys.
     * 
     */
    metafields: (Metafield | null)[]
    /** A note that is associated with the cart. For example, the note can be a personalized message to the buyer. */
    note: (Scalars['String'] | null)
    /** The total number of items in the cart. */
    totalQuantity: Scalars['Int']
    /** The date and time when the cart was updated. */
    updatedAt: Scalars['DateTime']
    __typename: 'Cart'
}
// ...
export interface CartBuyerIdentity {
    /** The country where the buyer is located. */
    countryCode: (CountryCode | null)
    /** The customer account associated with the cart. */
    customer: (Customer | null)
    /**
     * An ordered set of delivery addresses tied to the buyer that is interacting with the cart.
     * The rank of the preferences is determined by the order of the addresses in the array. Preferences
     * can be used to populate relevant fields in the checkout flow.
     * 
     */
    deliveryAddressPreferences: DeliveryAddress[]
    /** The email address of the buyer that is interacting with the cart. */
    email: (Scalars['String'] | null)
    /** The phone number of the buyer that is interacting with the cart. */
    phone: (Scalars['String'] | null)
    /**
     * A set of wallet preferences tied to the buyer that is interacting with the cart.
     * Preferences can be used to populate relevant payment fields in the checkout flow.
     * 
     */
    walletPreferences: Scalars['String'][]
    __typename: 'CartBuyerIdentity'
}
// ...
export interface CartCost {
    /** The estimated amount, before taxes and discounts, for the customer to pay at checkout. The checkout charge amount doesn't include any deferred payments that'll be paid at a later date. If the cart has no deferred payments, then the checkout charge amount is equivalent to `subtotalAmount`. */
    checkoutChargeAmount: MoneyV2
    /** The amount, before taxes and cart-level discounts, for the customer to pay. */
    subtotalAmount: MoneyV2
    /** Whether the subtotal amount is estimated. */
    subtotalAmountEstimated: Scalars['Boolean']
    /** The total amount for the customer to pay. */
    totalAmount: MoneyV2
    /** Whether the total amount is estimated. */
    totalAmountEstimated: Scalars['Boolean']
    /** The duty amount for the customer to pay at checkout. */
    totalDutyAmount: (MoneyV2 | null)
    /** Whether the total duty amount is estimated. */
    totalDutyAmountEstimated: Scalars['Boolean']
    /** The tax amount for the customer to pay at checkout. */
    totalTaxAmount: (MoneyV2 | null)
    /** Whether the total tax amount is estimated. */
    totalTaxAmountEstimated: Scalars['Boolean']
    __typename: 'CartCost'
}
// ...
export interface CartDeliveryGroupConnection {
    /** A list of edges. */
    edges: CartDeliveryGroupEdge[]
    /** A list of the nodes contained in CartDeliveryGroupEdge. */
    nodes: CartDeliveryGroup[]
    /** Information to aid in pagination. */
    pageInfo: PageInfo
    __typename: 'CartDeliveryGroupConnection'
}
// ...
export type CartDiscountAllocation = (CartAutomaticDiscountAllocation | CartCodeDiscountAllocation | CartCustomDiscountAllocation) & { __isUnion?: true }


/** The discount codes applied to the cart. */
export interface CartDiscountCode {
    /** Whether the discount code is applicable to the cart's current contents. */
    applicable: Scalars['Boolean']
    /** The code for the discount. */
    code: Scalars['String']
    __typename: 'CartDiscountCode'
}
// ...
export interface CartEstimatedCost {
    /** The estimated amount, before taxes and discounts, for the customer to pay at checkout. The checkout charge amount doesn't include any deferred payments that'll be paid at a later date. If the cart has no deferred payments, then the checkout charge amount is equivalent to`subtotal_amount`. */
    checkoutChargeAmount: MoneyV2
    /** The estimated amount, before taxes and discounts, for the customer to pay. */
    subtotalAmount: MoneyV2
    /** The estimated total amount for the customer to pay. */
    totalAmount: MoneyV2
    /** The estimated duty amount for the customer to pay at checkout. */
    totalDutyAmount: (MoneyV2 | null)
    /** The estimated tax amount for the customer to pay at checkout. */
    totalTaxAmount: (MoneyV2 | null)
    __typename: 'CartEstimatedCost'
}


/** Represents information about the merchandise in the cart. */
export interface CartLine {
    /** An attribute associated with the cart line. */
    attribute: (Attribute | null)
    /** The attributes associated with the cart line. Attributes are represented as key-value pairs. */
    attributes: Attribute[]
    /** The cost of the merchandise that the buyer will pay for at checkout. The costs are subject to change and changes will be reflected at checkout. */
    cost: CartLineCost
    /** The discounts that have been applied to the cart line. */
    discountAllocations: CartDiscountAllocation[]
    /**
     * @deprecated Use `cost` instead.
     * The estimated cost of the merchandise that the buyer will pay for at checkout. The estimated costs are subject to change and changes will be reflected at checkout.
     */
    estimatedCost: CartLineEstimatedCost
    /** A globally-unique ID. */
    id: Scalars['ID']
    /** The merchandise that the buyer intends to purchase. */
    merchandise: Merchandise
    /** The quantity of the merchandise that the customer intends to purchase. */
    quantity: Scalars['Int']
    /** The selling plan associated with the cart line and the effect that each selling plan has on variants when they're purchased. */
    sellingPlanAllocation: (SellingPlanAllocation | null)
    __typename: 'CartLine'
}


/** The cost of the merchandise line that the buyer will pay at checkout. */
export interface CartLineCost {
    /** The amount of the merchandise line. */
    amountPerQuantity: MoneyV2
    /** The compare at amount of the merchandise line. */
    compareAtAmountPerQuantity: (MoneyV2 | null)
    /** The cost of the merchandise line before line-level discounts. */
    subtotalAmount: MoneyV2
    /** The total cost of the merchandise line. */
    totalAmount: MoneyV2
    __typename: 'CartLineCost'
}


/** The estimated cost of the merchandise line that the buyer will pay at checkout. */
export interface CartLineEstimatedCost {
    /** The amount of the merchandise line. */
    amount: MoneyV2
    /** The compare at amount of the merchandise line. */
    compareAtAmount: (MoneyV2 | null)
    /** The estimated cost of the merchandise line before discounts. */
    subtotalAmount: MoneyV2
    /** The estimated total cost of the merchandise line. */
    totalAmount: MoneyV2
    __typename: 'CartLineEstimatedCost'
}
// ...
export type DeliveryAddress = (MailingAddress) & { __isUnion?: true }
// ...
export type Merchandise = (ProductVariant) & { __isUnion?: true }


/**
 * Metafields represent custom metadata attached to a resource. Metafields can be sorted into namespaces and are
 * comprised of keys, values, and value types.
 * 
 */
export interface Metafield {
    /** The date and time when the storefront metafield was created. */
    createdAt: Scalars['DateTime']
    /** The description of a metafield. */
    description: (Scalars['String'] | null)
    /** A globally-unique ID. */
    id: Scalars['ID']
    /** The unique identifier for the metafield within its namespace. */
    key: Scalars['String']
    /** The container for a group of metafields that the metafield is associated with. */
    namespace: Scalars['String']
    /** The type of resource that the metafield is attached to. */
    parentResource: MetafieldParentResource
    /** Returns a reference object if the metafield's type is a resource reference. */
    reference: (MetafieldReference | null)
    /** A list of reference objects if the metafield's type is a resource reference list. */
    references: (MetafieldReferenceConnection | null)
    /**
     * The type name of the metafield.
     * Refer to the list of [supported types](https://shopify.dev/apps/metafields/definitions/types).
     * 
     */
    type: Scalars['String']
    /** The date and time when the metafield was last updated. */
    updatedAt: Scalars['DateTime']
    /** The data stored in the metafield. Always stored as a string, regardless of the metafield's type. */
    value: Scalars['String']
    __typename: 'Metafield'
}
// ...
export interface AttributeInput {
/** Key or name of the attribute. */
key?: Scalars['String'],
/** Value of the attribute. */
value?: Scalars['String']}
// ...
export interface CartLineInput {
/** An array of key-value pairs that contains additional information about the merchandise line. */
attributes?: (AttributeInput[] | null),
/** The quantity of the merchandise. */
quantity?: (Scalars['Int'] | null),
/** The ID of the merchandise that the buyer intends to purchase. */
merchandiseId?: Scalars['ID'],
/** The ID of the selling plan that the merchandise is being purchased with. */
sellingPlanId?: (Scalars['ID'] | null)}


/** The input fields to update a line item on a cart. */
export interface CartLineUpdateInput {
/** The ID of the merchandise line. */
id?: Scalars['ID'],
/** The quantity of the line item. */
quantity?: (Scalars['Int'] | null),
/** The ID of the merchandise for the line item. */
merchandiseId?: (Scalars['ID'] | null),
/** An array of key-value pairs that contains additional information about the merchandise line. */
attributes?: (AttributeInput[] | null),
/** The ID of the selling plan that the merchandise is being purchased with. */
sellingPlanId?: (Scalars['ID'] | null)}